Move labels model to virtool.labels.models and update related usage
